+++ projects/demo-maps/src/app/components/example-layer-description/example-layer-description.component.ts
@@ -3,8 +3,7 @@ import { Component, Input, OnInit } from '@angular/core';
 @Component({
     selector: 'app-example-layer-description',
     templateUrl: './example-layer-description.component.html',
-    styleUrls: ['./example-layer-description.component.scss'],
-    standalone: true
+    styleUrls: ['./example-layer-description.component.scss']
 })
 export class ExampleLayerDescriptionComponent implements OnInit {
   @Input() layer;