+++ projects/layer-control/src/lib/utils/obj-type.pipe.ts
@@ -1,8 +1,7 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
-    name: 'type', pure: false,
-    standalone: true
+    name: 'type', pure: false
 })
 export class ObjTypePipe implements PipeTransform {
   transform(observations: any[], type: string): any[] {
@@ -15,8 +14,7 @@ export class ObjTypePipe implements PipeTransform {
  * e.g. usage @for(item of items | itemsfilter: callbackfn)
  */
 @Pipe({
-    name: 'itemsfilter', pure: false,
-    standalone: true
+    name: 'itemsfilter', pure: false
 })
 export class ItemsFilterPipe implements PipeTransform {
   transform(items: any[], callbackfn?: (value: any, index: number, array: any[]) => boolean): any[] {